requirements.txt
ultralytics>=8.1.34
paddlepaddle>=2.6.1
paddleocr==2.7.2
fastapi>=0.110.0
pydantic>=2.6.4
uvicorn>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
